symlink packages on install
