+++ scripts/bootstrap.js
@@ -11,44 +11,8 @@
  *     there on an "external" install.
  */
 const child_process = require('child_process');
-const fs = require('fs');
+const { allPackages, getPackageDeps } = require('./packages');
 
-// hardcoded for now
-const deps = {
-  'packages/dashboard': [
-    'packages/react-components',
-    'packages/api-client',
-    'packages/rmf-auth',
-    'packages/rmf-models',
-  ],
-  'packages/react-components': ['packages/api-client', 'packages/rmf-models'],
-  'packages/reporting': ['packages/react-components', 'packages/rmf-auth'],
-  'packages/api-client': ['packages/rmf-models'],
-};
-
-function getDeps(pkg) {
-  const recur = (pkg, cur) => {
-    if (deps[pkg]) {
-      deps[pkg].forEach((p) => {
-        recur(p, cur);
-      });
-    }
-    cur.add(pkg);
-    return cur;
-  };
-  return recur(pkg, new Set());
-}
-
-const allPackages = [
-  'packages/ros2-bridge',
-  'packages/react-components',
-  'packages/rmf-auth',
-  'packages/reporting',
-  'packages/dashboard',
-  'packages/api-server',
-  'packages/api-client',
-  'packages/rmf-models',
-];
 const scope = process.argv.length > 2 ? process.argv.slice(2) : allPackages;
 const verb = process.env['CI'] ? 'ci' : 'install';
 
@@ -56,23 +20,10 @@ const result = child_process.spawnSync('npm', [verb], { stdio: 'inherit', cwd: `
 if (result.status !== 0) {
   process.exit(result.status);
 }
-allPackages.forEach((pkg) => {
-  const packageJson = JSON.parse(fs.readFileSync(`${pkg}/package.json`));
-  const pkgName = packageJson.name;
-  try {
-    fs.unlinkSync(`node_modules/${pkgName}`);
-  } catch (e) {
-    if (e.code !== 'ENOENT') {
-      throw e;
-    }
-  }
-  fs.symlinkSync(`../${pkg}`, `node_modules/${pkgName}`);
-  console.log(`symlinked ${pkgName}`);
-});
 
 const targets = new Set();
 scope.forEach((pkg) => {
-  getDeps(pkg).forEach((p) => targets.add(p));
+  getPackageDeps(pkg).forEach((p) => targets.add(p));
 });
 targets.forEach((pkg) => {
   const cwd = `${__dirname}/../${pkg}`;
